requirements.txt
streamlit==1.37.1
pandas==2.2.2
openpyxl==3.1.5
streamlit-authenticator==0.4.2
PyYAML==6.0.2